+++ R/ts_graphs.R
@@ -27,7 +27,7 @@
 #' * `grid,<x,y,both>` specifies which gri lines to show.  Default is "both"
 #' * `norange` turns off date range selector.
 #' @param splitcols,stepcols,hidecols  String or list of data series to show on a second y axis, to be shown as step plots, or to be hidden.
-#' Can also be `TRUE` in which case first series in the data is affected. Can also be a semicolon separated single string with mutiple series.
+#' Can also be `TRUE` in which case first series in the data is used Can also be a semicolon separated single string with mutiple series.
 #' @param hilightcols String or list of data series to plot in different style than other series.
 #' @param hilightwidth (Default: 2) relative width of series specified in `hilightcols`
 #' @param hilightstyle (Default: solid).  Line style of series specified in `hilightcols`.
@@ -225,7 +225,7 @@ fgts_dygraph<-function(indata,title="",xlab="",ylab="",roller="default",bg_opts=
   # Misc date stuff
   col_date_list <- c(dt_colnames[['date']])
   alldts <- sort(unique(indt[[1]]))
-  dtlimits <- range(alldts)
+  dtlimits <- as.Date(as.double(range(alldts)))  # s/b fixed in next data.table release
   dtsrange_todisplay <- c(alldts[length(alldts)*dtstartfrac+1], max(alldts))
   if(nchar(dtwindow)>1) dtsrange_todisplay <- gendtstr(dtwindow,rtn="list")